+++ server/middleware/mockServer.js
@@ -139,6 +139,7 @@ module.exports = async (ctx, next) => {
   // let hostname = ctx.hostname;
   // let config = yapi.WEBCONFIG;
   let path = ctx.path;
+  let header = ctx.request.header;
 
   if (path.indexOf('/mock/') !== 0) {
     if (next) await next();
@@ -150,7 +151,10 @@ module.exports = async (ctx, next) => {
   paths.splice(0, 3);
   path = '/' + paths.join('/');
 
-  ctx.set('Access-Control-Allow-Origin', '*');
+  ctx.set('Access-Control-Allow-Origin', header.origin);
+  ctx.set('Access-Control-Allow-Credentials', true);
+
+  // ctx.set('Access-Control-Allow-Origin', '*');
 
   if (!projectId) {
     return (ctx.body = yapi.commons.resReturn(null, 400, 'projectId不能为空'));